BSOTranslation bug fix

+++ gip.core.datamodel/DbScripts/iPlus/dbsync_2021-06-18_13-16_aagincic.sql
@@ -1,4 +1,4 @@
-﻿IF EXISTS (SELECT * FROM sysobjects WHERE type = 'V' AND name = 'VBTranslationView')
+IF EXISTS (SELECT * FROM sysobjects WHERE type = 'V' AND name = 'VBTranslationView')
 	BEGIN
 		DROP  View dbo.[VBTranslationView]
 	END
